Unblock hard-test build: cast s_order/penalty_order to f64; fix test rng borrow ordering
- src/families/survival_predict.rs: annotate prev_total: f64 (ambiguous {float})
- src/terms/basis.rs: cast m / thin_plate_penalty_order(dimension) to f64
- tests/survival_multi_z_margpreserve_hard.rs: pre-bind rng.next_unit()/range() before &mut rng
- tests/coefficient_groups_hierarchical_hard.rs: &self receiver on EvaluateForTest helper
- tests/penalty_prior_mean_hard.rs: prefix unused rng bindings

Surfaces 3 real failures (left intact, as intended):
- covariance_autoderiv t01 + reduction auto_derivation_shape_reductions: classifier marks
  IID/orthogonal score columns as Full instead of Diagonal (off-diag threshold too strict)
- fit_hard truth_neglog_minimised: a -5% perturbation undercuts neglog at population truth

+++ src/terms/basis.rs
@@ -19861,7 +19861,7 @@ fn thin_plate_kernel_triplet_from_scaled_distance(
         3 => Ok((-scaled_distance, -1.0, 0.0)),
         _ => polyharmonic_kernel_triplet(
             scaled_distance,
-            thin_plate_penalty_order(dimension),
+            thin_plate_penalty_order(dimension) as f64,
             dimension,
         ),
     }
